Swapping LMGTFY link for GitHub instructions link
The Imagine Math team culture focuses on kindness and helpfulness.
Having a LMGTFY link in our instructions sends the wrong message
to the Globant team and comes off as unprofessional.  The link
has been updated with the GitHub help link to properly reflect
our current style of documentation and tone.